tee: split cli part of tee.rs
